requirements.txt
beautifulsoup4==4.14.3
dtaidistance==2.3.13
enlighten==1.14.1
fastdtw==0.3.4
fitz==0.0.1.dev2
html_to_markdown==2.24.5
httpx==0.28.1
langchain_text_splitters==1.1.0
langextract==1.1.1
loguru==0.7.3
matplotlib==3.10.8
numpy==2.4.2
omegaconf-argparse==1.0.1
openai==2.17.0
openai-agents==0.6.4
pandas==3.0.0
Pillow==12.1.0
playwright==1.58.0
pymupdf==1.26.7
pymupdf-layout==1.26.6
pymupdf4llm==0.2.2
requests==2.31.0
scikit_learn==1.8.0
scipy==1.17.0
sentence_transformers==5.1.2
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1